+++ src/MigrationHistory.php
@@ -3,10 +3,12 @@
 namespace Dynart\Micro\Entities;
 
 use Dynart\Micro\Entities\Attribute\Column;
+use Dynart\Micro\Entities\Attribute\Table;
 
 /**
  * The record of one applied migration
  */
+#[Table(name: 'migration_history')]
 class MigrationHistory extends Entity {
 
     protected static string $eventName = 'migration_history';